Add files via upload
Hello.   I implemented the changes that you suggested and was able to run my program.   Thanks for your help.   I'm having issues with the last 2 sections of the project and I'm down to 1 hour 20 min on GPU timer, so I need assistance.   Prior to uploading the file the file to GitHub, the error for both sections (Class Prediction and Sanity Checking) stated that the file directory was not found, but I believe the file directory listed was correct 'flower_data/train/5/image_05151.jpg'.    Could you check my code for the these sections and let me know how to fix the error?    I appreciate it.